requirements.txt
tensorflow
datasets
numpy
PyYAML
